Added domain and range specification to mrel.
Can't assume that we will always have representative pairs for
everything in the range.  Really only need everything at the upper end
of the range, but this is still an issue.  Still not getting results I
expect with construct_explicit_relation when the computation domain is
smaller than the data index space.