Proved computation graphs always compute a value, and are deterministic.
This is subject to the requirement that the graphs be well-formed.

Next step is to show that any for loop will give rise to a well-formed
computation graph that computes the same value. After that, I will
need to show that if we have a good \delta to hand, then the graph
will be the same.