Get graphOf to halfway reasonable state prior to radical rework.
In particular, I now think that it's impossible to get the graph
framework to cope with the difference between local and global scopes
and, worse, nested local scopes. So, instead I'll assume one memory
map, and rely on higher levels to translate variable names apart from
one another.

The initialisation of local variables in for- and par- loop iterations
can be explicitly done with assignments that are generated by the
semantics as it executes.