Fix imap_irrelevance result for action graphs.
This result says that if the actions ignore their iteration numbers,
then the result of evaluating a whole graph of such actions can ignore
them too.  The fix is to remove an unnecessary assumption about the
renaming being injective because if it isn't, then the imap is just
the identity function on the graph, meaning that it is indeed
irrelevant.