Progress with graphOf correctness proof, but...
Revision to model is going to be needed to record the fact that the
construction of the graph depends on "reads" of memory that aren't
apparent only in the Assign constructor. In particular, there are
reads that happen in the course of evaluating the guards of
if-statements and the like.