Get PseudoC semantics to build with new handling of arrays.
The semantics gets a deal simpler, because AssignVar has disappeared,
and so too has the dexpr type. The only thing the latter loses us is
that index calculation for array assignments can't be interleaved with
data-reads. Instead, the address of an lvalue is calculated all at
once, when the RHS is ready to go.

Lots else does not build.