Prove "weak" rule for FOR loops.
The previous rule (now called "strong") is the equivalent of the Hoare
rule for showing termination and correctness of loops, while this one
assumes termination.  In fact, PseudoC's ForLoops are guaranteed to
terminate, but they aren't sure to terminate successfully, so this is
the notion of "termination" these rules cover.